# Conflicts:
#	mock/usecase-integration/src/main/kotlin/app/meetacy/backend/mock/integration/users/CreateUserUsecase.kt
#	usecase/src/main/kotlin/app/meetacy/backend/usecase/friends/add/AddFriendUsecase.kt

+++ database/usecase-integration/src/main/kotlin/app/meetacy/backend/database/integration/notifications/GetNotificationsUsecase.kt
@@ -22,7 +22,7 @@ class DatabaseGetNotificationStorage(db: Database) : GetNotificationsUsecase.Sto
         count: Int
     ): List<GetNotificationsUsecase.NotificationFromStorage> =
         notificationsTable
-            .getNotifications(userId, offset.toInt(), count)
+            .getNotifications(userId, offset, count)
             .map(DatabaseNotification::mapToUsecase)
 
 }
